# docs/api-docs.yaml
openapi: 3.0.1
info:
  title: API Service Request
  description: >-
    This API is responsible for managing the information from the technical test
    database.
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
  /users:
    post:
      tags:
        - users
      summary: Insert a new user in the database.
      requestBody:
        content:
          aplication/json:
            schema:
              $ref: '#/components/schemas/BodyUserPost'
        required: true
      responses:
        '200':
          description: success
          content:
            aplication/json:
              schema:
                $ref: '#/components/schemas/userCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
    put:
      tags:
        - users
      summary: Insert a new user in the database.
      parameters:
        - $ref: '#/components/parameters/token'
      requestBody:
        content:
          aplication/json:
            schema:
              $ref: '#/components/schemas/BodyTechnicalPost'
        required: true
      responses:
        '200':
          description: success
          content:
            aplication/json:
              schema:
                $ref: '#/components/schemas/technicalCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /users/{idUser}:
    get:
      tags:
        - users
      summary: get one user in the database.
      parameters:
        - $ref: '#/components/parameters/token'
        - name: idUser
          in: path
          description: User id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: success
          content:
            aplication/json:
              schema:
                $ref: '#/components/schemas/technicalCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /users/technical:
    post:
      tags:
        - users
      summary: Insert a new technical in the database.
      requestBody:
        content:
          aplication/json:
            schema:
              $ref: '#/components/schemas/BodyUserPost'
        required: true
      responses:
        '200':
          description: success
          content:
            aplication/json:
              schema:
                $ref: '#/components/schemas/technicalCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /users/admin:
    post:
      tags:
        - users
      summary: Insert a new admin in the database.
      requestBody:
        content:
          aplication/json:
            schema:
              $ref: '#/components/schemas/BodyUserPost'
        required: true
      responses:
        '200':
          description: success
          content:
            aplication/json:
              schema:
                $ref: '#/components/schemas/userCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /users/add-service:
    put:
      tags:
        - users
      summary: Insert a new service del Technical in the database.
      parameters:
        - $ref: '#/components/parameters/token'
      requestBody:
        content:
          aplication/json:
            schema:
              $ref: '#/components/schemas/serviceAddBody'
      responses:
        '200':
          description: success
          content:
            aplication/json:
              schema:
                $ref: '#/components/schemas/serviceAdd'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
  /users/delete-service:
    put:
      tags:
        - users
      summary: Delete a service of technical in the database.
      parameters:
        - $ref: '#/components/parameters/token'
      requestBody:
        content:
          aplication/json:
            schema:
              $ref: '#/components/schemas/serviceAddBody'
      responses:
        '200':
          description: success
          content:
            aplication/json:
              schema:
                $ref: '#/components/schemas/serviceAdd'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  responses:
    Unauthorized:
      description: (Unauthorized) no authorization when calling the service
    NotFound:
      description: (NotFound) the information was not found
    BadRequest:
      description: (BadRequest) the data sent is incorrect
    ServerError:
      description: (ServerError) the data sent is incorrect
  parameters:
    token:
      name: token
      in: header
      description: Authentication token
      required: true
  schemas:
    BodyUserPost:
      type: object
      properties:
        firstName:
          type: string
          description: User First name
        lastName:
          type: string
          description: User last name
        email:
          type: string
          description: User email
        password:
          type: string
          description: User password
        telephone:
          type: string
          description: User telephone
    BodyTechnicalPost:
      type: object
      properties:
        firstName:
          type: string
          description: User First name
        lastName:
          type: string
          description: User last name
        email:
          type: string
          description: User email
        password:
          type: string
          description: User password
        telephone:
          type: string
          description: User telephone
        services:
          type: array
          items:
            type: string
          description: technical Services
    userCreated:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          type: object
          properties:
            id:
              type: integer
              description: User ID
            firstName:
              type: string
              description: User First name
            lastName:
              type: string
              description: User last name
            email:
              type: string
              description: User email
            password:
              type: string
              description: User password
            status:
              type: string
              description: User status
            telephone:
              type: string
              description: User telephone
          example:
            firstName: Cliente2
            lastName: lastName
            email: cliente2@gmail.com
            telephone: '+584121574000'
            status: Active
            id: 7
    technicalCreated:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          type: object
          properties:
            id:
              type: integer
              description: User ID
            firstName:
              type: string
              description: User First name
            lastName:
              type: string
              description: User last name
            email:
              type: string
              description: User email
            password:
              type: string
              description: User password
            status:
              type: string
              description: User status
            telephone:
              type: string
              description: User telephone
            services:
              type: array
              items:
                type: string
              description: technical Services
          example:
            firstName: Cliente2
            lastName: lastName
            email: cliente2@gmail.com
            telephone: '+584121574000'
            status: Active
            id: 7
            services:
              - Maintanance
              - Support
    serviceAddBody:
      type: object
      properties:
        service:
          type: string
    serviceAdd:
      type: object
      properties:
        success:
          type: boolean
          example: true
        services:
          type: array
          items:
            type: string
          description: technical Services
          example:
            - Maintanance
            - Support
...